kiro/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
sentence-transformers==2.2.2

python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.24.1
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
transformers==4.35.0
torch==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-telegram-bot==20.3
deep-translator